Fetches data for admin overview dashboard
Updates the admin overview dashboard to fetch data from the server instead of using mock data.

This change improves data accuracy and allows for a more dynamic and up-to-date representation of key metrics.
It also adds loading states to the dashboard to improve the user experience while data is being fetched.

Adds `database_export.json` to `.gitignore`.
Updates the Faculty icon in the sidebar.